% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R, R/guide_legend_subcat.R,
%   R/scale_colour_subcat.R
\docType{data}
\name{ggsubcat-ggproto}
\alias{ggsubcat-ggproto}
\alias{GuideLegendSubcat}
\alias{ScaleSubcat}
\title{Base ggproto classes for ggsubcat}
\description{
Base ggproto classes for ggsubcat
}
\section{Methodology}{

As nicely described in the book \emph{GGPLOT2: Elegant Graphics for Data
Analysis}, the graph building theory comes with the data and definition of
selected aesthetics. After opting to add the desired geoms, ggplot2 system
maps the aesthetics to their corresponding scales, which project the data
values to the visual information (e.g. x/y to position, color-by variable to
color). Lastly, Guides are generated to help reading the graph so that people
can map the visual information back to data values (e.g. axes for reading x/y,
legend for knowing what each color represents).

Following the streamline, each aesthetic leads to a scale and finally yields
a guide. For the sake of tidy visualization, guides coming from the same
aesthetics defined by the same data variables can be merged (e.g.
with \code{aes(colour = sample, shape = sample)}, you'll see colored point
each of a different shape in one legend).

We designed ggsubcat to scale the sub-categorical information, made up by two
(aesthetic) variables, into one set of color and legend. The core trick
behind would be to allow the merging of two guides from different variables.
We first made the scale to be aware of two aesthetics that come in pair, and
have the mapping information cached in the scale. Then we modified the
hashables to be checked in the guides to avoid the previously described
blockage and turn the merging mechanism to check the sub-categorical mapping
previously cached, which makes sense.
}

\section{Guides}{

Object \code{GuideLegendSubcat} is provided for building sub-categorical
legends from \code{ScaleSubcat}. It inherits the original ggplot2 guide
\code{GuideLegend} and is modified with the following properties:
\itemize{
\item{\code{params} - Fields of \code{customTitle}, \code{subcatMap} are
added.}
\item{\code{hashables} - Are modified to use \code{title}, \code{subcatMap},
and \code{name} to compute the hash. Only in this way can we merge the
internally resulted two guides for e.g. color and color.group with different
aesthetic variables. As a result, this guide can never be merged with other
guides.}
}

The following methods are modified:
\itemize{
\item{\code{train()} - Avoids auto title generation that the original
\code{GuideLegend} does and in turn do not generate general title. And
extracts the \code{subcatMap} from the scale.}
\item{\code{setup_params()} - Computes number of breaks, nrow and ncol for
each main category.}
\item{\code{build_subtitles()} - A new method that overrides the place where
the original \code{build_title()} is called. It generates title grobs for
each main category.}
\item{\code{build_decor()} - Build the key grobs only for one main category
at a time.}
\item{\code{draw()} - Splits the key table by main categories and build
grob table for each main category, i.e. draw a "legend" for each main
category, and combine them into one grob table.}
\item{\code{measure_grobs()} and \code{arrange_layout()} - For the grob
table of each main category, measures the width and height of the grob and
arrange the element layout.}
}
}

\section{Scales}{

Object \code{ScaleSubcat} is provided for sub-categorical color scaling.

It inherits from ggplot2's original scale \code{ScaleDiscrete} and is
modified with the following properties:
\itemize{
\item{params - A list of additional parameters and caches to be used for
forming the palette.}
}

And also modified with the following methods:
\itemize{
\item{train_df() - Checks the pair of aesthetics (e.g. \code{colour} and
\code{colour.group}), identifies the correct sub-categorical mapping, and
predefines the palette for the categories in e.g. \code{colour.group}.}
\item{palette() - Uses the pre-defined parameters to form the colors being
used.}
}
}

\keyword{datasets}
