% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_legend_subcat.R
\name{guide_legend_subcat}
\alias{guide_legend_subcat}
\title{Sub-categorical legend guide}
\usage{
guide_legend_subcat(
  title = waiver(),
  theme = NULL,
  position = NULL,
  direction = NULL,
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  reverse = FALSE,
  order = 0,
  ...
)
}
\arguments{
\item{title}{A character string or expression indicating an overall title of
guide. If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2]{waiver}()}), the name of the scale object or the name
specified in \code{\link[ggplot2]{labs}()} is used for the title.}

\item{theme}{A \code{\link[ggplot2]{theme}} object to style the guide
individually or differently from the plot's theme settings. The \code{theme}
argument in the guide overrides, and is combined with, the plot's theme.}

\item{position}{A character string indicating where the legend should be
placed relative to the plot panels.}

\item{direction}{A character string indicating the direction of the guide.
One of \code{"horizontal"} or \code{"vertical"}.}

\item{override.aes}{A list specifying aesthetic parameters of legend key. See
details and examples. TODO, anyway to avoid Duplicated \code{override.aes}
warning? Since at graph building time, two copies of the same guide are
generated for e.g. \code{color} and \code{color.group} and then merged.}

\item{nrow, ncol}{The desired number of rows and column of legends
respectively. TODO, this applies to inidividual group for now but ideally we
need a better scheme for overall arranging of all groups.}

\item{reverse}{logical. If \code{TRUE} the order of legends is reversed.
TODO, make \code{reverse.inner} and \code{reverse.outer} separetely.}

\item{order}{Positive integer less than 99 that specifies the order of this
guide among multiple guides. This controls the order in which multiple guides
are displayed, not the contents of the guide itself. If 0 (default), the
order is determined by a secret algorithm.}

\item{...}{Ignored}
}
\description{
Legend type guide shows key (i.e. geoms) mapped onto values. Different than
a ggplot2 original legend guide, keys are split by parent categories and then
parent category names are added as sub-titles. Integration with other scales
is not possible.
}
\examples{
library(ggplot2)
ggplot(subcatPoints, aes(x, y, color = sub, color.group = main)) +
    geom_point() +
    scale_color_subcat(guide = guide_legend_subcat(
        position = "bottom"
    ))
}
