% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colors.R
\name{makeSubcatCol}
\alias{makeSubcatCol}
\title{Generate colors for sub-categories}
\usage{
makeSubcatCol(
  main,
  sub,
  mainColors = NULL,
  lMin = 20,
  lMax = 80,
  cMin = NULL,
  cMax = NULL
)
}
\arguments{
\item{main}{A factor/vector representing the variable with the main
categories}

\item{sub}{A factor/vector representing the variable with the sub-categories.}

\item{mainColors}{A named/unnamed vector/list of colors for the main
categories. Colors should be presented with a string color code that R can
understand. For example \code{"red"} or \code{"#FF0000"}. Default \code{NULL}
use evenly spaced hues (ggplot2 default discrete colors).}

\item{lMin, lMax}{Minimum and maximum luminance for the shaded colors, must be
between [0, 100] with min less then max. Default is \code{20} and \code{80}.}

\item{cMin, cMax}{Minimum and maximum chroma for the HCL color space, must be
between [0, 100] with min less than max. Default is \code{NULL} for both.}
}
\value{
list object. Each element is named by the categories in \code{main}.
Each element is a named vector of colors for the sub-categories mapped to the
corresponding main category.
}
\description{
This function first maps sub-categories to corresponding main categories,
determines the colors for main categories, and then generates shaded colors
for sub-categories belonging to each main category.
}
\details{
The shading process mainly works by tuning the luminance of a given color
in HCL space. The range of acceptable luminance is determined by \code{lMin}
and \code{lMax}. Since the given color has its luminance, the function
decides whether to increase or decrease the luminance based on the given
range. If the luminance of the given color is larger than the mean of the
range, the function will decrease the luminance from the given color to
\code{lMin}. Otherwise, it will increase the luminance from the given color
to \code{lMax}. It by default increases or decreases the luminance by 20.
When more colors are needed, the function will evenly space the luminance.

The chroma can be tuned in the meantime. If none of \code{cMin} or
\code{cMax} is set (all \code{NULL}), the chroma of the given color will be
retained. If only one of \code{cMin} or \code{cMax} is given, the function
will use the given value for the colors of all sub-categories. If both are
given, the function will evenly space the chroma between the two values, from
max to min.
}
\examples{
makeSubcatCol(
    main = rep(letters[1:3], 3),
    sub = paste0(rep(letters[1:3], 3), 1:9)
)
}
