% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_subcat.R
\name{scale_colour_subcat}
\alias{scale_colour_subcat}
\alias{scale_color_subcat}
\alias{scale_fill_subcat}
\title{Sub-categorical colour scales}
\usage{
scale_colour_subcat(lMin = 20, lMax = 80, cMin = NULL, cMax = NULL, ...)

scale_fill_subcat(lMin = lMin, lMax = lMax, cMin = cMin, cMax = cMax, ...)
}
\arguments{
\item{lMin, lMax}{Minimum and maximum luminance for the shaded colors, must be
between [0, 100] with min less then max. Default is \code{20} and \code{80}.}

\item{cMin, cMax}{Minimum and maximum chroma for the HCL color space, must be
between [0, 100] with min less than max. Default is \code{NULL} for both.}

\item{...}{other arguments that constructs a discrete scale. TODO.}
}
\description{
The sub-categorical colour scale. Work with one of the pairs of aesthetics
specified: \code{colour} and \code{colour.group} or \code{fill} and
\code{fill.group}. A level in the variable that \code{colour} specifies must
map to one and only one level in the variable that \code{colour.group}
specifies. The same rule applies to the pair \code{fill} and
\code{fill.group}.
}
\examples{
library(ggplot2)
ggplot(subcatPoints, aes(x, y, color = sub, color.group = main)) +
    geom_point() +
    scale_colour_subcat()

ggplot(subcatPoints, aes(sub, y, fill = sub, fill.group = main)) +
   geom_bar(stat = "identity") +
   scale_fill_subcat()
}
